backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.9.1
python-dotenv==1.0.1
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.27
aiosqlite==0.19.0
alembic==1.13.1

# LLM and AI dependencies
openai==1.12.0
anthropic==0.18.1
transformers==4.37.2
torch==2.2.0
sentence-transformers==2.3.1

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1
python-pptx==0.6.23
openpyxl==3.1.2

# Utilities
python-slugify==8.0.4
tenacity==8.2.3
httpx==0.26.0
rich==13.7.0
loguru==0.7.2

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0
